# swagger/swagger.yaml
swagger: '2.0'
info:
  description: >-
    Current API provides access to AuthService, which is a part of SecureFiles
    project, which gives you powerfull tool to upload your files in completely
    secure mode and share it with people you trust to see this absolutely secret
    files.
  version: 1.0.2
  title: Auth Service
  termsOfService: 'http://service-name-here/terms/'
  contact:
    email: apiteam@authservice.com
host: 'localhost:8080'
basePath: /v0
produces:
  - application/xml
  - application/json
...
schemes:
  - http
paths:
  /users:
    post:
      consumes:
        - application/xml
        - application/json
      tags:
        - User
      summary: Create new user
      description: This endpoint allows to add new user.
      operationId: addUser
      parameters:
        - in: body
          name: user
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/UserCredentials'
      responses:
        '201':
          description: user added
          schema:
            $ref: '#/definitions/User'
        '400':
          description: invalid input
          schema:
            $ref: '#/definitions/Error'
      security:
        - authservice_auth:
            - 'write:users'
            - 'read:users'
    get:
      summary: List users by filter
      description: This endpoint allows to get list of users searched by filter criteria.
      operationId: listUsersByFilter
      tags:
        - User
      parameters:
        - name: filter
          in: query
          description: object with parameters by which to filter users
          required: false
          schema:
              $ref: '#/definitions/UsersFilter'
      responses:
        '200':
          description: a list of filtered users
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
  '/users/{userId}':
    get:
      summary: Get user by Id
      description: This endpoint allows to get user by Id.
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: user Id to get by
          required: true
          type: integer
          format: int64
          minimum: 1
      tags:
        - User
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Update user
      description: Update existing user
      operationId: updateUserById
      parameters:
        - name: userId
          in: path
          description: Id of user that will be updated
          required: true
          type: integer
          format: int64
          minimum: 1
        - in: body
          name: body
          description: Updated user
          required: true
          schema:
            $ref: '#/definitions/User'
      tags:
        - User
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete user from database
      description: Deletes User
      operationId: deleteUserById
      parameters:
        - name: userId
          in: path
          description: Id of user that will be deleted
          required: true
          type: integer
          format: int64
          minimum: 1
      tags:
        - User
      responses:
        '200':
          description: Removed
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
      security:
        - authservice_auth:
            - 'write:users'
            - 'read:users'
  '/users/{userId}/pass':
    put:
      summary: Update user including password
      description: Update existing user including password
      operationId: updateUserCredentials
      parameters:
        - name: userId
          in: path
          description: Id of user that will be updated
          required: true
          type: integer
          format: int64
          minimum: 1
        - in: body
          name: userCredentials
          description: Updated user credentials
          required: true
          schema:
            $ref: '#/definitions/UserCredentials'
      tags:
        - User
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
  '/users/{userId}/role':
    put:
      summary: Update user role
      description: Change existing user role
      operationId: updateUserRole
      parameters:
        - name: userId
          in: path
          description: Id of user whose role to be updated
          required: true
          type: integer
          format: int64
          minimum: 1
        - in: body
          name: user
          description: Updated user role
          required: true
          schema:
            $ref: '#/definitions/User'
      tags:
        - User
      responses:
        '200':
          description: role updated
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
  '/users/{userId}/activate':
    patch:
      summary: Activate user
      description: Activates user
      operationId: activateUser
      parameters:
        - name: userId
          in: path
          description: Id of user to be activated
          required: true
          type: integer
          format: int64
          minimum: 1
      tags:
        - User
      responses:
        '200':
          description: user activated
          schema:
            $ref: '#/definitions/User'
        '400':
          description: failed to activate user
          schema:
            $ref: '#/definitions/Error'
  /users/list/{offset}/{limit}:
    get:
      tags:
        - User
      summary: Return list of users per page
      description: This endpoint returns a list of users. The list is limited by offset and number of results on page
      operationId: listUsersByOffsetAndPageLimit
      parameters:
        - name: offset
          in: path
          description: Position in pagination
          required: true
          type: integer
          format: int32
          minimum: 0
        - name: limit
          in: path
          description: Number of items to retrieve (5 min, 50 max)
          required: true
          type: integer
          format: int32
          minimum: 5
          maximum: 50
      responses:
        '200':
          description: a list of users
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '400':
          description: invalid input
          schema:
            $ref: '#/definitions/Error'
      security:
        - authservice_auth:
            - 'write:users'
            - 'read:users'
...
  '/stat/size/{userId}':
    get:
      tags:
        - Statistics
      summary: Get size of all uploaded files for this user
      description: ''
      operationId: statSize
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of user to get statistics for"
          required: true
          type: "string"
      responses:
        '200':
          description: successfull operation
          schema:
            type: string
        '400':
          description: Bad request
        default:
          description: invalid input
          schema:
            $ref: '#/definitions/Error'
securityDefinitions:
  authservice_auth:
    type: oauth2
    authorizationUrl: 'http://projectname.com'
    flow: implicit
    scopes:
      'write:users': modify users
      'read:users': read users
      'write:files': modify files
      'read:files': read files
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      firstname:
        type: string
        minLength: 2
        maxLength: 30
      lastname:
        type: string
        minLength: 2
        maxLength: 30
      email:
        type: string
      role:
        $ref: '#/definitions/Role'
    xml:
      name: User
  UserCredentials:
    type: object
    required:
      - password
      - email
    properties:
      id:
        type: integer
        format: int64
      firstname:
        type: string
        minLength: 2
        maxLength: 30
      lastname:
        type: string
        minLength: 2
        maxLength: 30
      email:
        type: string
      password:
        type: string
        minLength: 4
        maxLength: 20
      role:
        $ref: '#/definitions/Role'
    xml:
      name: UserCredentials
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string
    xml:
      name: Error
...
  Role:
    type: string
    description: User athorization level
    enum:
      - ADMIN
      - TECH_SUPP
      - USER
  UsersFilter:
    type: object
    properties:
      offset:
        type: integer
        format: int32
        minimum: 0
        description: Position in pagination
      limit:
        type: integer
        format: int32
        minimum: 5
        maximum: 50
        description:  Number of items to retrieve (5 min, 50 max)
      order:
        type: string
        enum: [ASC, DESC]
        description: Order in which to sort list of users. By defauld sorted by ID
      sortField:
        type: string
        enum: [userId, email, firstName, lastName]
        description: Field in User by which to sort list of users
      role:
        $ref: '#/definitions/Role'
      userStatus:
        type: string
        enum: [ACTIVE, DELETED]
        default: ACTIVE
        description: Status of user
      passwordStatus:
        type: string
        enum: [ACTIVE, EXPIRED]
        description: Status of user's password for ACTIVE users. For DELETED users password status is ignored.
    xml:
      name: UsersFilter
...